VictoryScreen code reorg
